Move testing logic to remoted module
