+++ tests/integration/test_remoted/test_agent_communication/test_shared_configuration.py
@@ -2,16 +2,13 @@
 # Created by Wazuh, Inc. <info@wazuh.com>.
 # This program is free software; you can redistribute it and/or modify it under the terms of GPLv2
 import os
-import subprocess
 from time import sleep
 
 import pytest
 import wazuh_testing.tools.agent_simulator as ag
 from wazuh_testing import UDP, TCP, TCP_UDP
-from wazuh_testing import remote as rd
-from wazuh_testing.tools import LOG_FILE_PATH
+from wazuh_testing.remote import check_push_shared_config
 from wazuh_testing.tools.configuration import load_wazuh_configurations
-from wazuh_testing.tools.monitoring import FileMonitor
 
 # Marks
 
@@ -43,7 +40,6 @@ agent_info = {
 
 configurations = load_wazuh_configurations(configurations_path, __name__, params=parameters, metadata=metadata)
 config_ids = [x['PROTOCOL'] for x in parameters]
-wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
 
 
 # fixtures
@@ -51,69 +47,6 @@ wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
 def get_configuration(request):
     """Get configurations from the module."""
     return request.param
-
-
-def check_push_shared_config(protocol):
-    """Allow to check if the manager sends the shared configuration to agents through remoted.
-
-    First, check if the default group configuration file is completely pushed (up message, configuration
-    and close message). Then add the agent to a new group and check if the new configuration is pushed.
-    Also it checks that the same config isn't pushed two times.
-
-    Args:
-        protocol (str): It can be UDP or TCP.
-
-    Raises:
-        TimeoutError: If agent does not receive the manager ACK message in the expected time.
-    """
-
-    # Create agent and sender object with default parameters
-    agent = ag.Agent(**agent_info)
-
-    # Sleep to avoid ConnectionRefusedError
-    sleep(1)
-
-    sender = ag.Sender(agent_info['manager_address'], protocol=protocol)
-
-    # Activate receives_messages modules in simulated agent.
-    agent.set_module_status('receive_messages', 'enabled')
-
-    # Run injector with only receive messages module enabled
-    injector = ag.Injector(sender, agent)
-    try:
-        injector.run()
-
-        # Wait until remoted has loaded the new agent key
-        rd.wait_to_remoted_key_update(wazuh_log_monitor)
-
-        # Send the start-up message
-        sender.send_event(agent.startup_msg)
-        sender.send_event(agent.keep_alive_event)
-
-        # Check up file (push start) message
-        rd.check_agent_received_message(agent.rcv_msg_queue, r'#!-up file \w+ merged.mg', timeout=10,
-                                        error_message="initial up file message not received")
-
-        # Check agent.conf message
-        rd.check_agent_received_message(agent.rcv_msg_queue, '#default', timeout=10,
-                                        error_message="agent.conf message not received")
-        # Check close file (push end) message
-        rd.check_agent_received_message(agent.rcv_msg_queue, 'close', timeout=10,
-                                        error_message="initial close message not received")
-        sender.send_event(agent.keep_alive_event)
-        # Check that push message doesn't appear again
-        with pytest.raises(TimeoutError):
-            rd.check_agent_received_message(agent.rcv_msg_queue, r'#!-up file \w+ merged.mg', timeout=5)
-            raise AssertionError("Same shared configuration pushed twice!")
-
-        # Add agent to group and check if the configuration is pushed.
-        subprocess.run(["/var/ossec/bin/agent_groups", "-q", "-a", "-i", agent.id, "-g", "testing_group"])
-        sender.send_event(agent.keep_alive_event)
-        rd.check_agent_received_message(agent.rcv_msg_queue, '#!-up file .* merged.mg', timeout=10,
-                                        error_message="New group shared config not received")
-
-    finally:
-        injector.stop_receive()
 
 
 def test_push_shared_config(get_configuration, configure_environment, restart_remoted, create_agent_group):
@@ -126,4 +59,8 @@ def test_push_shared_config(get_configuration, configure_environment, restart_re
     protocols = get_configuration['metadata']['protocol']
 
     for protocol in protocols.split(","):
-        check_push_shared_config(protocol)
+        agent = ag.Agent(**agent_info)
+        # Sleep to avoid ConnectionRefusedError
+        sleep(1)
+        sender = ag.Sender(agent_info['manager_address'], protocol=protocol)
+        check_push_shared_config(protocol, agent, sender)
